--- src/greedyPathfind.hpp
@@ -1,26 +1,0 @@
-#include "vec2.hpp"
-#include "internalMap.hpp"
-#include "logger.hpp"
-#include "util.h"
-
-
-class pathfinder {
-    Vec2 targetCoord;
-    
-    internalMap* mapInstance;
-
-
-    //should take a reference to the instantiated map, will be done later for convenience
-    pathfinder();
-
-    char heuristic();
-    void newTarget(Vec2 newTgt);
-    void newMovePattern();
-    void dumpMovePattern();
-
-    void wallRedirect();
-
-
-
-
-};
