+++ src/pathfinderClasses.cpp
@@ -1,4 +1,4 @@
-#include "greedyPathfind.hpp"
+#include "pathfinderClasses.hpp"
 
 //this is for a hound, fox will need some minor changes
 char pathfinder::heuristic(){
@@ -6,14 +6,25 @@ char pathfinder::heuristic(){
     double cDist = linDist(mapInstance->currentPos(),targetCoord);
     
     //if straight ahead is better, go there
-    if(cDist > linDist(mapInstance->currentPos()+mapInstance->getHeadingVector(),targetCoord)){
-        return 'F';
-    }
+    
 
     //if right is better, rotate, the next turn will sort movement out
     Vec2 tempHeading = mapInstance->getHeadingVector(); 
     Vec2 cPos = mapInstance->currentPos();
     
+    if(!mapInstance->priorVisit(cPos)){
+        pastVisits++;
+    }
+
+
+    if(cDist > linDist(cPos+tempHeading,targetCoord)){
+        return 'F';
+    }
+
+    if(pastVisits == 15){
+        //enter unstuck mode, will be a different function, has to determine when to xero out the trigger variable
+        //return unstuck();
+    }
     ninetyClockwise(tempHeading);
     if(cDist > linDist(tempHeading+cPos,targetCoord)){
         if(!mapInstance->isWall(cPos)){
@@ -22,19 +33,26 @@ char pathfinder::heuristic(){
     }
 
     ninetyClockwise(tempHeading);
-    if(cDist > linDist(tempHeading+targetCoord,targetCoord)){
+    if(cDist > linDist(tempHeading+cPos,targetCoord)){
         if(!mapInstance->isWall(cPos)){
             return 'R';
         }
     }
 
     ninetyClockwise(tempHeading);
-    if(cDist > linDist(tempHeading+,targetCoord)){
+    if(cDist > linDist(tempHeading+cPos,targetCoord)){
         if(!mapInstance->isWall(cPos)){
-            return ""
+            return 'R';
         }    
     }
 
+
+
+
 }
 
+
+char pathfinder::unstuck(){
+    
+}
